backend/requirements.txt
-i https://pypi.org/simple
asgiref==3.2.10; python_version >= '3.5'
cachetools==4.1.1; python_version ~= '3.5'
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
django-cors-headers==3.5.0
django-storages==1.10.1
django==3.1.2
djangorestframework-simplejwt==4.4.0
djangorestframework==3.12.1
google-api-core[grpc]==1.22.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
google-auth==1.22.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
google-cloud-core==1.4.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
google-cloud-secret-manager==2.0.0
google-cloud-storage==1.31.2
google-crc32c==1.0.0; python_version >= '3.5'
google-resumable-media==1.1.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
googleapis-common-protos[grpc]==1.52.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
grpc-google-iam-v1==0.12.3
grpcio==1.32.0
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
libcst==0.3.13; python_version >= '3.6'
mypy-extensions==0.4.3
pillow==7.2.0
proto-plus==1.10.1
protobuf==3.13.0
psycopg2==2.8.6
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyjwt==1.7.1
pytz==2020.1
pyyaml==5.3.1
requests==2.24.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
rsa==4.6; python_version >= '3.5'
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.1; python_version >= '3.5'
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.25.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
